Extracted Extracted methods for interaction with SearchNotesFragment from HomeScreenRobot to HomeSearchNoteRobot
